refactor: rename --> AppInsights
