+++ packages/salesforcedx-utils-vscode/src/telemetry/reporters/appInsights.ts
