+++ packages/salesforcedx-utils-vscode/src/telemetry/telemetry.ts
@@ -15,7 +15,7 @@ import {
 import { SfdxSettingsService } from '../settings';
 import { disableCLITelemetry, isCLITelemetryAllowed } from './cliConfiguration';
 import { TelemetryReporter } from './interfaces/telemetryReporter';
-import { AppInsights } from './reporters/appInsightsReporter';
+import { AppInsights } from './reporters/appInsights';
 import { LogStream as LogStream } from './reporters/logStream';
 import { TelemetryFileReporter } from './reporters/telemetryFileReporter';
 import { LogStreamConfig } from './logStreamConfig';
